pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "psychai"
version = "0.1.0"
description = "A generic framework for training and evaluating models for psychology research"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Jingfeng(Craig) Zhang", email = "jz44@illinois.edu"},
]
maintainers = [
    {name = "Jingfeng(Craig) Zhang", email = "jz44@illinois.edu"},
]
keywords = [
    "machine learning",
    "natural language processing",
    "transformers",
    "training",
    "evaluation",
    "pytorch",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.12.0",
    "transformers>=4.20.0",
    "datasets>=2.0.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "tqdm>=4.60.0",
    "peft>=0.3.0",
    "accelerate>=0.20.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
unsloth = [
    "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git",
    "trl>=0.7.0",
]
audio = [
    "librosa>=0.10.0",
]
vision = [
    "Pillow>=10.0.0",
    "torchvision>=0.16.0",
]
vision-timm = [
    "timm>=0.9.0",
]
vision-hf = [
    "transformers>=4.38.0",
    "accelerate>=0.24.0",
]
vision-clip = [
    "open_clip_torch>=2.24.0",
]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "flake8>=4.0", 
    "pytest-cov>=3.0",
    "mypy>=0.900",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
    "myst-parser>=0.17.0",
]

[project.urls]
Homepage = "https://github.com/cracrazjf/PsychAI"
Documentation = "https://github.com/cracrazjf/PsychAI#readme"
Repository = "https://github.com/cracrazjf/PsychAI.git"
"Bug Tracker" = "https://github.com/cracrazjf/PsychAI/issues"

# [project.scripts]
# Scripts will be added when CLI modules are implemented
# psychai-train = "psychai.scripts.train:main"
# psychai-evaluate = "psychai.scripts.evaluate:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["psychai*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=PsychAI",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true